+++ tests/setup_env.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Setup environment variables to point to a sandbox 
+# Setup environment variables to point to a sandbox
 
 TEST_ROOT=$(cd `dirname "${BASH_SOURCE[0]}"` && pwd)
 TEST_DATA=$TEST_ROOT/data