+++ trigger/conf/global_settings.py
@@ -67,9 +67,11 @@ SUPPORTED_VENDORS = (
     'cisco',
     'citrix',
     'dell',
+    'f5',
     'force10',
     'foundry',
     'juniper',
+    'mrv',
     'netscreen',
     'paloalto',
 )
@@ -89,9 +91,11 @@ VENDOR_MAP = {
     'CISCO SYSTEMS': 'cisco',
     'CITRIX': 'citrix',
     'DELL': 'dell',
+    'F5 NETWORKS': 'f5',
     'FORCE10': 'force10',
     'FOUNDRY': 'foundry',
     'JUNIPER': 'juniper',
+    'MRV': 'mrv',
     'NETSCREEN TECHNOLOGIES': 'netscreen',
 }
 
@@ -105,15 +109,18 @@ SUPPORTED_PLATFORMS = {
     'cisco': ['ROUTER', 'SWITCH'],
     'citrix': ['SWITCH'],                         # Assumed to be NetScalers
     'dell': ['SWITCH'],
+    'f5': ['LOAD BALANCING', 'SWITCH'],
     'force10': ['ROUTER', 'SWITCH'],
     'foundry': ['ROUTER', 'SWITCH'],
     'juniper': ['FIREWALL', 'ROUTER', 'SWITCH'],  # Any devices running Junos
+    'mrv': ['CONSOLE SERVER', 'SWITCH'],
     'netscreen': ['FIREWALL'],                    # Pre-Juniper NetScreens
     'paloalto': ['FIREWALL'],
 }
 
 # The tuple of support device types
-SUPPORTED_TYPES = ('FIREWALL', 'DWDM', 'LOAD BALANCER', 'ROUTER', 'SWITCH')
+SUPPORTED_TYPES = ('CONSOLE SERVER', 'FIREWALL', 'DWDM', 'LOAD BALANCING',
+                   'ROUTER', 'SWITCH')
 
 # A mapping of of vendor names to the default device type for each in the
 # event that a device object is created and the deviceType attribute isn't set
@@ -126,9 +133,11 @@ DEFAULT_TYPES = {
     'citrix': 'SWITCH',
     'cisco': 'ROUTER',
     'dell': 'SWITCH',
+    'f5': 'LOAD BALANCING',
     'force10': 'ROUTER',
     'foundry': 'SWITCH',
     'juniper': 'ROUTER',
+    'mrv': 'CONSOLE SERVER',
     'netscreen': 'FIREWALL',
     'paloalto': 'FIREWALL',
 }
@@ -261,11 +270,14 @@ JUNIPER_FULL_COMMIT_FIELDS = {
 # Specially-defined, per-vendor prompt patterns. If a vendor isn't defined here,
 # try to use IOSLIKE_PROMPT_PAT or fallback to DEFAULT_PROMPT_PAT.
 PROMPT_PATTERNS = {
-    'aruba': r'\(\S+\)(?: \(\S+\))?\s?#',
+    'aruba': r'\(\S+\)(?: \(\S+\))?\s?#', # ArubaOS 6.1
+    #'aruba': r'\S+(?: \(\S+\))?\s?#\s', # ArubaOS 6.2
     'citrix': r'\sDone\n$',
+    'f5': r'(?:\S+\@)?\S+(?:\(.*\))\(tmos\)#\s{1,2}\r?$',
     'juniper': r'\S+\@\S+(?:\>|#)\s$',
-    'paloalto': r'\r\n\S+(?:\>|#)\s?',
+    'mrv': r'\r\n?.*(?:\:\d{1})?\s\>\>?$',
     'netscreen': r'(\w+?:|)[\w().-]*\(?([\w.-])?\)?\s*->\s*$',
+    'paloalto': r'\r\n\S+(?:\>|#)\s?',
 }
 
 # When a pattern is not explicitly defined for a vendor, this is what we'll try