+++ trigger/contrib/commando/plugins/__init__.py
@@ -1,1 +1,8 @@
 """Plugins for the Commando API built around Trigger."""
+"""
+trigger.contrib.commando.plugins
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+This package provides facilities for running commands on devices using the CLI.
+Plugins for the Commando API built around Trigger.
+"""