+++ trigger/contrib/commando/plugins/show_version.py
@@ -1,7 +1,7 @@
 
+from twisted.python import log
 from trigger.contrib.commando import CommandoApplication
 from trigger.utils import xmltodict, strip_juniper_namespace
-from twisted.python import log
 
 task_name = 'show_version'