+++ trigger/tacacsrc.py
@@ -21,8 +21,8 @@ from time import strftime, localtime
 import os
 import pwd
 import sys
-from trigger.conf import settings
 from twisted.python import log
+from trigger.conf import settings
 
 # Exports
 __all__ = ('get_device_password', 'prompt_credentials', 'convert_tacacsrc',
@@ -371,7 +371,9 @@ class Tacacsrc(object):
         creds[realm] = prompt_credentials(realm, user)
         log.msg('setting self.creds_updated flag', debug=True)
         self.creds_updated = True
-        print '\nCredentials updated for user: %r, device/realm: %r.' % (user, realm)
+        new_user = creds[realm].username
+        print '\nCredentials updated for user: %r, device/realm: %r.' % \
+              (new_user, realm)
 
     def _encrypt_old(self, s):
         """Encodes using the old method. Adds a newline for you."""
@@ -393,6 +395,7 @@ class Tacacsrc(object):
 
     def _read_file_old(self):
         """Read old style file and return the raw data."""
+        self._update_perms()
         with open(self.file_name, 'r') as f:
             return f.readlines()